+++ src/selectly.js
@@ -1,16 +1,14 @@
-import { Select as ReactARIASelect } from 'react-aria'
 import Select from './Select'
 import Trigger from './Trigger'
 import Menu from './Menu'
+import Option from './Option'
 import buildOptionsLookup from './utils/build-options-lookup'
 import getCurrentOptions from './utils/get-current-options'
 import getToggledValues from './utils/get-toggled-values'
 import getAllValues from './utils/get-all-values'
 import isOptionSelected from './utils/is-option-selected'
 
-const { Option } = ReactARIASelect
-
-export const utils = {
+const utils = {
   buildOptionsLookup,
   getCurrentOptions,
   getToggledValues,
@@ -22,5 +20,6 @@ export {
   Select,
   Trigger,
   Menu,
-  Option
+  Option,
+  utils
 }